Updated SPC export logic
Updated SPC exporting to scape out a portion of the program block and cram as many samples into the SPC as it can. This required moving the note lengths table as well and updating the program to point to the new location. Also note, exported SPCs can play but can't be reimported for some reason.