Fixed resetting the playback state
Loading a ROM used to clear the playback state, but it stopped working with one of the previous commits. This change changes the functionality to assume `close_rom` means to close the the current ROM and/or any currently loaded SPCs and will correctly reset the playback state in either case to prevent crashes.