Make fixed-width font Consolas and fix sizing of tabs
It doesn't have all the same serifs, but it looks a lot nicer, comes with
Windows since Vista, and it behaves a lot nicer with the scaling code I used.

There's clearly something funny going on with NONCLIENTMETRICS. I got a value of
8 for iPaddedBorderWidth when using the Windows SDK in MSVC, but 0 when using
MinGW-W64. I'm not sure if this is due to a difference in the size of a struct
or something else.