Dynamically determine song pointer table location
Previously, it was assumed that the song pointer table was at a fixed location in the ROM. However, with CoilSnake 4.2 beta 1 and newer, all of the song packs, including the engine containing the song pointer table, can be relocated.

In order to open ROMs compiled by this version of CoilSnake, when loading all of the packs we now check if the current block is the main engine block, and if so, save the address in the file of the song pointer table based on that block's address.